Variable Time-step Implementation
Main game loop now uses variable time-step instead of fixed time-step.

+++ Src/AIManager.cpp
@@ -22,7 +22,7 @@ AIManager::~AIManager()
 	delete m_MapGraph;
 }
 
-void AIManager::Update( void* aiManagerInst, Entity** entities, int startIndex, int numEntities, float deltaTime )
+void AIManager::Update( void* aiManagerInst, Entity** entities, int startIndex, int numEntities, double deltaTime )
 {
 	// Cast to AIManager pointer for us to use.
 	AIManager* aiManager = (AIManager*)aiManagerInst;
