+++ Src/AIManager.h
@@ -35,7 +35,7 @@ public:
 	// @param startIndex array index to start from
 	// @param numEntity How many entities are in the array
 	// @param deltaTime time passed since the last frame update
-	static void Update( void* aiManagerInst, Entity** entities, int startIndex, int numEntities, float deltaTime );
+	static void Update( void* aiManagerInst, Entity** entities, int startIndex, int numEntities, double deltaTime );
 
 private:
 	// Function to compute for cost of travelling between two given segments.
