+++ Src/Player.cpp
@@ -4,7 +4,7 @@ Player::Player( char renderCharacter )
 	:	Entity(renderCharacter)
 {}
 
-void Player::Update( float deltaTime )
+void Player::Update( double deltaTime )
 {
 
 }
